+++ app/components/uchi/field/belongs_to.rb
@@ -3,7 +3,7 @@
 module Uchi
   class Field
     class BelongsTo < Field
-      DEFAULT_COLLECTION_QUERY = ->(query:) { query }.freeze
+      DEFAULT_COLLECTION_QUERY = ->(context:, query:) { query }.freeze
 
       module Helpers
         def associated_record
@@ -28,7 +28,7 @@ module Uchi
           return nil if model.nil?
 
           repository_class = Uchi::Repository.for_model(model)
-          repository_class.new
+          repository_class.new(context: repository.context)
         end
 
         def label_for_associated_record
@@ -50,7 +50,7 @@ module Uchi
             return nil if model.nil?
 
             repository_class = Uchi::Repository.for_model(model)
-            repository_class.new
+            repository_class.new(context: repository.context)
           end
         end