+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - All procs are now called with keyword arguments. This means procs configured for `collection_query`, `searchable`, `sortable` will need to have their signature changed (from eg `lambda { |query, term| }` to `lambda { |query:, term:| }`).
 
+### Added
+
+- `Uchi::Context` with useful details about what's currently happening.
+- `collection_query`, `searchable`, `sortable` procs now receive a `context` argument with the current `Uchi::Context`.
+
 ### Changed
 
 - Code from `Uchi::ApplicationController` has been moved to `Uchi::Controller`. This provides a clean `Uchi::ApplicationController` for consumer applications to reopen and modify.